Move xfel.util.jungfrau to serialtbx.detector
